Reorganize specs for submissions
